🐛(compose) fix "port already taken" errors when starting docker compose
We have changed the project's name from "impress" to "docs" but haven't
replace all occurrences of impress in the project because we want to be
careful of the consequences on deployments.

The name of the docker compose project was different for the pyling make
target which was causing this bug. Let's rename if without waiting.